Bugfix: Force EEG data imported from .mat file to double

+++ toolbox/io/in_data_mat.m
@@ -103,7 +103,7 @@ DataMat = repmat(DataMat, [nbEpoch, 1]);
 % Process each epoch
 BaseComment = DataMat(1).Comment;
 for i = 1:nbEpoch
-    DataMat(i).F = FileData(:,:,i);
+    DataMat(i).F = double(FileData(:,:,i));
     % Add indice number for multiple epochs
     if (nbEpoch > 1)
         DataMat(i).Comment = sprintf('%s #%d', BaseComment, i);
